moved error-boundary.tsx
