+++ src/error-boundary.tsx
@@ -10,6 +10,9 @@ interface State {
   errorInfo?: ErrorInfo;
 }
 
+/**
+ * Error Boundary for internal use.
+ */
 export class ErrorBoundary extends Component<Props, State> {
   public static getDerivedStateFromError(error: Error): State {
     // Update state so the next render will show the fallback UI.
@@ -30,21 +33,19 @@ export class ErrorBoundary extends Component<Props, State> {
 
     if (state.hasError) {
       return (
-        <>
-          <div className="ctw-conditions ctw-stacked">
-            <div className="ctw-heading-container">
-              <div className="ctw-title">The Bad News</div>
-            </div>
-            <div className="ctw-body-container">
-              <div className="ctw-space-y-3">
-                <div className="ctw-title-container">
-                  <div className="ctw-title">Error:</div>
-                  <p>{state.error?.message || "Unknown Error"}</p>
-                </div>
+        <div className="ctw-conditions ctw-stacked">
+          <div className="ctw-heading-container">
+            <div className="ctw-title">The Bad News</div>
+          </div>
+          <div className="ctw-body-container">
+            <div className="ctw-space-y-3">
+              <div className="ctw-title-container">
+                <div className="ctw-title">Error:</div>
+                <p>{state.error?.message || "Unknown"}</p>
               </div>
             </div>
           </div>
-        </>
+        </div>
       );
     }
     return props.children;
